Created function to establish connection
Created a function to establish connectoin with mysql server and updated the documentation. Tested it and it works.
Removed test.py since I no longer need to have it and my connectoin correctly works.
Also, added a version control in the documentation to give information about past and current state of development of the software